MMO-7: Rename file to match extension loader and fix information command
